Implement Jobs update api

+++ app/api/admin/jobs/[id]/route.ts
@@ -1,7 +1,50 @@
+import { authOptions } from "@/lib/auth";
 import connectDB from "@/lib/db";
 import { Job } from "@/models/Job";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 
+// PATCH: Update job details or status
+export async function PATCH(
+  req: Request,
+  { params }: { params: { id: string } },
+) {
+  try {
+    const { id } = await params;
+    await connectDB();
+    const session = await getServerSession(authOptions);
+
+    if (!session || session.user.role !== "admin") {
+      return NextResponse.json(
+        { success: false, message: "Unauthorized" },
+        { status: 403 },
+      );
+    }
+
+    const body = await req.json();
+    const updatedJob = await Job.findByIdAndUpdate(
+      id,
+      { $set: body },
+      { new: true }, // Returns the updated document
+    );
+
+    if (!updatedJob) {
+      return NextResponse.json(
+        { success: false, message: "Job not found" },
+        { status: 404 },
+      );
+    }
+
+    return NextResponse.json({ success: true, data: updatedJob });
+  } catch (error) {
+    return NextResponse.json(
+      { success: false, message: "Update failed" },
+      { status: 500 },
+    );
+  }
+}
+
+// DELETE API Route
 export async function DELETE(
   req: Request,
   { params }: { params: { id: string } },
@@ -9,7 +52,23 @@ export async function DELETE(
   try {
     const { id } = await params;
     await connectDB();
-    await Job.findByIdAndDelete(id);
+    const session = await getServerSession(authOptions);
+
+    if (!session || session.user.role !== "admin") {
+      return NextResponse.json(
+        { success: false, message: "Unauthorized" },
+        { status: 403 },
+      );
+    }
+
+    const deletedJob = await Job.findByIdAndDelete(id);
+
+    if (!deletedJob) {
+      return NextResponse.json(
+        { success: false, message: "Job not found" },
+        { status: 404 },
+      );
+    }
     return NextResponse.json({ success: true, message: "Job deleted" });
   } catch (error) {
     return NextResponse.json({ message: "Delete failed" }, { status: 500 });
